+++ CHANGELOG.md
@@ -24,6 +24,8 @@ All notable changes to this project will be documented in this file.
 -   Updated working directory
 -   Updated nginx container from v2.1.0 to v3.0.0
 -   Removed general-cli:3.1.1
+-   Removed eslint-loader v4.0.2
+-   Removed snyk v1.502.0
 
 ### Added
 
@@ -38,6 +40,7 @@ All notable changes to this project will be documented in this file.
 -   Added a info_xdebug.php file
 -   Added php in front of the phpcs phpcbf commands in gulp files
 -   Added light-cli:1.0.0
+-   Added eslint-webpack-plugin v2.5.2
 
 ## [8.2.0] - 2020-08-24